+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 Meteor A-Frame Package Change Log
 ==================================
 
+## 0.3.0 - March 23, 2016
+
+* Upgrade to aframe 0.2.0-rc1
+* Update FlickrJump example
+* Added `aEvent` helper
+* Added new primitives
+  * `aBox`
+  * `aCircle`
+  * `aColladaModel`
+  * `aCone`
+  * `aCursor`
+  * `aObjModel`
+  * `aRing`
+* Deprecated primitives are still supported:
+  * `aCube`
+  * `aModel`
+
 ## 0.2.0 - March 17, 2016
 
 * Attributes are now set via helpers, not in the `onRendered` callback